requirements.txt
aiohttp==3.7.4.post0
amqp==2.1.4
ansible==3.4.0
ansible-base==2.10.9
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
billiard==3.5.0.5
blinker==1.4
cached-property==1.5.2
celery==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
decorator==4.2.1
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
falcon==1.4.1
Flask==1.1.2
idna==2.8
idna-ssl==1.1.0
importlib-metadata==3.10.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.2.2.post1
lxml==4.2.5
MarkupSafe==1.1.1
marshmallow==2.13.0
multidict==5.1.0
packaging==20.9
pycparser==2.20
pyformance==0.4
PyJWT==1.4.2
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.18.0
python-Levenshtein==0.12.0
python-mimeparse==1.6.0
python-redis-lock==3.2.0
pytz==2021.1
PyYAML==5.4.1
raven==6.6.0
redis==2.10.5
requests==2.21.0
requests-toolbelt==0.9.1
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.32.14
tabulate==0.7.7
twilio==6.10.5
typing-extensions==3.7.4.3
urllib3==1.24.3
vine==1.3.0
watchdog==2.0.2
webargs==1.5.1
Werkzeug==1.0.1
yarl==1.6.3
zeep==3.1.0
zipp==3.4.1
